perf(server): lighten shared test setup
