+++ apps/server/src/orpc/routes/bottleAliases/create.ts
@@ -9,7 +9,7 @@ import {
 import { logError } from "@peated/server/lib/log";
 import { procedure } from "@peated/server/orpc";
 import { requireMod } from "@peated/server/orpc/middleware";
-import { pushUniqueJob } from "@peated/server/worker/client";
+import { pushUniqueJob } from "@peated/server/worker/dispatch";
 import { z } from "zod";
 
 export default procedure
