+++ apps/server/src/orpc/routes/bottleSeries/update.ts
@@ -10,7 +10,7 @@ import {
 } from "@peated/server/schemas";
 import { serialize } from "@peated/server/serializers";
 import { BottleSeriesSerializer } from "@peated/server/serializers/bottleSeries";
-import { pushUniqueJob } from "@peated/server/worker/client";
+import { pushUniqueJob } from "@peated/server/worker/dispatch";
 import { and, eq, ne, sql } from "drizzle-orm";
 import { z } from "zod";
