+++ apps/server/src/orpc/routes/entities/aliases/delete.ts
@@ -1,7 +1,7 @@
 import { deleteEntityAlias } from "@peated/server/lib/entityAliases";
 import { procedure } from "@peated/server/orpc";
 import { requireMod } from "@peated/server/orpc/middleware";
-import { pushUniqueJob } from "@peated/server/worker/client";
+import { pushUniqueJob } from "@peated/server/worker/dispatch";
 import { z } from "zod";
 
 export default procedure
