+++ apps/server/src/orpc/routes/entities/references/create.ts
@@ -5,7 +5,7 @@ import { changes, entities, entityReferences } from "@peated/server/db/schema";
 import { getUserActorForDatabase } from "@peated/server/lib/actors";
 import { procedure } from "@peated/server/orpc";
 import { requireMod } from "@peated/server/orpc/middleware";
-import { pushUniqueJob } from "@peated/server/worker/client";
+import { pushUniqueJob } from "@peated/server/worker/dispatch";
 import { and, eq, isNull, sql } from "drizzle-orm";
 import { z } from "zod";
