+++ apps/server/src/worker/jobs/capturePriceImage.ts
@@ -3,7 +3,7 @@ import { db } from "@peated/server/db";
 import { storePrices } from "@peated/server/db/schema";
 import { logInfo, logTelemetryError } from "@peated/server/lib/log";
 import { compressAndResizeImage, storeFile } from "@peated/server/lib/uploads";
-import { pushUniqueJob } from "@peated/server/worker/client";
+import { pushUniqueJob } from "@peated/server/worker/dispatch";
 import type { JobContext, JobPayload } from "@peated/server/worker/types";
 import { eq } from "drizzle-orm";
 import { Readable } from "stream";
