+++ apps/server/src/worker/jobs/onBottleChange.ts
@@ -1,5 +1,5 @@
 import { db } from "@peated/server/db";
-import { pushUniqueJob, runJob } from "@peated/server/worker/client";
+import { pushUniqueJob, runJob } from "@peated/server/worker/dispatch";
 import { z } from "zod";
 import type { JobPayload } from "../types";
 import type { UpdateBottleStatsJobArgs } from "./updateBottleStats";
